File & variable name refactoring.
