+++ Buddies.Web/src/components/CreateProjectForm.tsx
@@ -9,7 +9,7 @@ import { CreateProjectRequest } from '../api/model/createProjectRequest';
 import { AuthState } from '../stores/authStore';
 import { makeThrottledSearch } from '../api/utils';
 import { SearchFunc } from '../api';
-import { scrollHandler } from './utils';
+import { scrollHandler } from './event-handlers';
 
 const count = 10;
