+++ Buddies.Web/src/components/InviteDialog.tsx
@@ -11,7 +11,7 @@ import TextField from '@mui/material/TextField';
 import { InviteUserRequest } from '../api/model/inviteUserRequest';
 import { makeThrottledSearch } from '../api/utils';
 import { SearchFunc } from '../api';
-import { scrollHandler } from './utils';
+import { scrollHandler } from './event-handlers';
 
 interface Props {
   open: boolean;
